+++ cerberus/__init__.py
@@ -1,6 +1,6 @@
 __all__ = ['client', 'user_auth', 'aws_auth']
 
-CLIENT_VERSION = '0.11.0'
+CLIENT_VERSION = '0.12.0'
 
 
 class CerberusClientException(Exception):